bugfix: add missing break; also move refund-IC creation to..
..invoiceCandidateFactory
Make refund contract doctypes #4083
